Move _ansi.h contents to sys/cdefs.h
Prepare to switch users from _ansi.h to sys/cdefs.h
